Input Validation

+++ js/controllers.js
@@ -24,13 +24,45 @@ app.controller('RootController', ['$scope', function($scope) { // Root
 
 app.controller('TarefasController', ['$scope', function($scope) { // Tarefas
   $scope.addTodo = function() {
-    $scope.todos.push({
-      'text': $scope.newTodo,
-      'cat': $scope.inCateg.cat,
-      'done': false
-    });
-    $scope.newTodo = '';
-    localStorage.setItem('todos', JSON.stringify($scope.todos));
+    var addToArrayToDo = true;
+    var addToArrayCat = true;
+
+    for (var i = 0; i < $scope.todos.length; i++) {
+      if ($scope.todos[i].text === $scope.newTodo) {
+        addToArrayToDo = false;
+      }
+    }
+    for (var x = 0; x < $scope.todos.length; x++) {
+      if ($scope.todos[x].cat === $scope.inCateg.cat) {
+        addToArrayCat = false;
+      }
+    }
+    if ((addToArrayToDo === false) && (addToArrayCat === false)) {
+      swal({
+        title: "Tarefa já existe!",
+        type: "error",
+        timer: 1000,
+        showConfirmButton: false
+      });
+    }
+    if ((addToArrayToDo === true) && (addToArrayCat === true)) {
+      $scope.todos.push({
+        'text': $scope.newTodo,
+        'cat': $scope.inCateg.cat,
+        'done': false
+      });
+      $scope.newTodo = '';
+      localStorage.setItem('todos', JSON.stringify($scope.todos));
+    }
+    if ((addToArrayToDo === false) && (addToArrayCat === true)) {
+      $scope.todos.push({
+        'text': $scope.newTodo,
+        'cat': $scope.inCateg.cat,
+        'done': false
+      });
+      $scope.newTodo = '';
+      localStorage.setItem('todos', JSON.stringify($scope.todos));
+    }
   };
 
   $scope.todoCheck = function(todo) {
